fed_llm/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
cmake=3.27.7=pypi_0
contourpy=1.1.1=pypi_0
cubinlinker-cu11=0.3.0.post1=pypi_0
cuda-python=11.8.2=pypi_0
cudf-cu11=23.10.0=pypi_0
cuml-cu11=23.10.0=pypi_0
cupy-cuda11x=12.2.0=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.4=pypi_0
dask=2023.9.2=pypi_0
dask-cuda=23.10.0=pypi_0
dask-cudf-cu11=23.10.0=pypi_0
distributed=2023.9.2=pypi_0
fastrlock=0.8.2=pypi_0
filelock=3.12.4=pypi_0
fonttools=4.43.1=pypi_0
fsspec=2023.9.2=pypi_0
fvcore=0.1.5.post20221221=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
iopath=0.1.10=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lit=17.0.3=pypi_0
llvmlite=0.40.1=pypi_0
locket=1.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
medpy=0.4.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.7=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numba=0.57.1=pypi_0
numpy=1.24.4=pypi_0
nvtx=0.2.8=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.8.1.78=pypi_0
openssl=3.0.11=h7f8727e_2
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
partd=1.4.1=pypi_0
pillow=10.1.0=pypi_0
pip=23.3=py39h06a4308_0
portalocker=2.8.2=pypi_0
protobuf=4.24.4=pypi_0
psutil=5.9.6=pypi_0
ptxcompiler-cu11=0.7.0.post1=pypi_0
pyarrow=12.0.1=pypi_0
pylibraft-cu11=23.10.0=pypi_0
pynvml=11.4.1=pypi_0
pyparsing=3.1.1=pypi_0
pyre-extensions=0.0.29=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
raft-dask-cu11=23.10.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
rmm-cu11=23.10.0=pypi_0
scikit-fuzzy=0.4.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
seaborn=0.13.0=pypi_0
setuptools=68.0.0=py39h06a4308_0
simpleitk=2.3.0=pypi_0
six=1.16.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.41.2=h5eee18b_0
submitit=1.5.0=pypi_0
surface-distance-based-measures=0.1=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tblib=2.0.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=pypi_0
torch=2.0.0+cu117=pypi_0
torchmetrics=0.10.3=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.15.0+cu117=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
treelite=3.9.1=pypi_0
treelite-runtime=3.9.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023c=h04d1e81_0
ucx-py-cu11=0.34.0=pypi_0
urllib3=2.0.7=pypi_0
wheel=0.41.2=py39h06a4308_0
xformers=0.0.19=pypi_0
xlrd=2.0.1=pypi_0
xlwt=1.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
zict=3.0.0=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
